rename students to student
